requirements.txt
Django
django-admin-interface
django-multiselectfield
django-json-widget
numpy
Pillow
questplus
pandas
requests
docker-compose
ipfml